feat: add APIEndpointName enum and StatsData type definitions
